14th May

+++ src/src/portfolio4/Class02.java
@@ -15,8 +15,8 @@ public class Class02 {
 	    Field[] fields = a.getClass().getFields();
 	    System.out.printf("There are %d fields\n", fields.length);
 	    for (Field f : fields) {
-	      System.out.printf("field name=%s type=%s value=%d\n", f.getName(),
-	          f.getType(), f.getInt(a));
+	      System.out.printf("field name=%s type=%s value=%f\n", f.getName(),
+	          f.getType(), f.getDouble(a));
 	    }
 
 	}
