+++ src/src/portfolio4/Class08.java
@@ -15,10 +15,10 @@ public class Class08 {
 	    System.out.printf("There are %d fields\n", fields.length);
 	    for (Field f : fields) {
 	      f.setAccessible(true);
-	      int x = f.getInt(a);
+	      double x = f.getDouble(a);
 	      x++;
-	      f.setInt(a, x);
-	      System.out.printf("field name=%s type=%s value=%d\n", f.getName(), f.getType(), f.getInt(a));
+	      f.setDouble(a, x);
+	      System.out.printf("field name=%s type=%s value=%f\n", f.getName(), f.getType(), f.getDouble(a));
 	    }
 
 	}
